week2-api/pyproject.toml
[project]
name = "week2-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.79.0",
    "python-dotenv>=1.2.1",
]
